% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad.times.R
\name{raad_process_files}
\alias{raad_process_files}
\title{taken from raadtools:::.processFiles}
\usage{
raad_process_files(dt, f, tr)
}
\arguments{
\item{dt}{date-time column (POSIXct)}

\item{f}{files (date, at least)}

\item{tr}{time resolution, see Details}
}
\description{
tr timeresolutions is something from 4hourly, 12hourly, 3hourly, 6hourly, daily, weekly, monthly, weekly3, 8daily, 8D
}
